Fixing include paths

+++ Source/Multiplayer/MPGameObjectContainer.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "Multiplayer\MPGameObjectContainer.h"
+#include "Multiplayer/MPGameObjectContainer.h"
 
 MPGameObjectContainer::MPGameObjectContainer() : _PlayerBulletSpeed(600), _AIBulletSpeed(400) {}
 
@@ -393,13 +393,13 @@ void MPGameObjectContainer::setAiCarFrequency()
 		_CarFrequency = 1.5f + 0.1f * (float)_Level;
 		break;
 	case 1:
-		_CarFrequency = 1.75f + 0.11f * std::powf((float)_Level, 1.3f);
+		_CarFrequency = 1.75f + 0.11f * std::pow((float)_Level, 1.3f);
 		break;
 	case 2:
-		_CarFrequency = 2.0f + 0.15f * std::powf((float)_Level, 1.3f);
+		_CarFrequency = 2.0f + 0.15f * std::pow((float)_Level, 1.3f);
 		break;
 	case 3:
-		_CarFrequency = 2.15f + 0.17 * std::powf((float)_Level, 1.45f);
+		_CarFrequency = 2.15f + 0.17 * std::pow((float)_Level, 1.45f);
 		break;
 	default:
 		break;
@@ -414,13 +414,13 @@ void MPGameObjectContainer::setBulletFrequency()
 		_BulletFrequency = 0.8f + 0.065f * (float)_Level;
 		break;
 	case 1:
-		_BulletFrequency = 1.2f + 0.08f * std::powf((float)_Level, 1.1f);
+		_BulletFrequency = 1.2f + 0.08f * std::pow((float)_Level, 1.1f);
 		break;
 	case 2:
-		_BulletFrequency = 1.2f + 1.0f * std::powf((float)_Level, 1.2f);
+		_BulletFrequency = 1.2f + 1.0f * std::pow((float)_Level, 1.2f);
 		break;
 	case 3:
-		_BulletFrequency = 1.4f + 1.0f * std::powf((float)_Level, 1.33f);
+		_BulletFrequency = 1.4f + 1.0f * std::pow((float)_Level, 1.33f);
 		break;
 	default:
 		break;
